requirements.txt
torch
torchvision
ultralytics
opencv-python
matplotlib
Pillow
requests
numpy
nltk
scipy
gensim
git+https://github.com/facebookresearch/segment-anything-2.git
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm